+++ packages/api/cms-api/src/dam/dam.module.ts
@@ -18,7 +18,6 @@ import { DamFileImage } from "./files/entities/file-image.entity";
 import { createFolderEntity, FolderInterface } from "./files/entities/folder.entity";
 import { FileImagesResolver } from "./files/file-image.resolver";
 import { FileLicensesResolver } from "./files/file-licenses.resolver";
-import { FileUploadService } from "./files/file-upload.service";
 import { FileValidationService } from "./files/file-validation.service";
 import { createFilesController } from "./files/files.controller";
 import { createFilesResolver } from "./files/files.resolver";
@@ -126,7 +125,6 @@ export class DamModule {
                 FileImagesResolver,
                 CalculateDominantImageColor,
                 FileValidationService,
-                FileUploadService,
                 PixelImageBlockTransformerService,
                 SvgImageBlockTransformerService,
                 DamVideoBlockTransformerService,
@@ -141,7 +139,6 @@ export class DamModule {
                 ImagesService,
                 ScaledImagesCacheService,
                 damConfigProvider,
-                FileUploadService,
                 PixelImageBlockTransformerService,
                 SvgImageBlockTransformerService,
                 DamVideoBlockTransformerService,