+++ packages/admin/admin-color-picker/CHANGELOG.md
@@ -201,6 +201,23 @@
     -   @comet/admin@7.0.0-beta.0
     -   @comet/admin-icons@7.0.0-beta.0
 
+## 6.17.1
+
+### Patch Changes
+
+-   @comet/admin@6.17.1
+-   @comet/admin-icons@6.17.1
+
+## 6.17.0
+
+### Patch Changes
+
+-   Updated dependencies [536e95c02]
+-   Updated dependencies [7ecc30eba]
+-   Updated dependencies [ec4685bf3]
+    -   @comet/admin@6.17.0
+    -   @comet/admin-icons@6.17.0
+
 ## 6.16.0
 
 ### Patch Changes